# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: "标题：ArchForge接口文档"
  description: "描述：文档说明"
  contact:
    name: "ArchForge"
  version: "版本号:1.0.0"
  license:
    name: "Apache-2.0"
    identifier: "Apache-2.0"
servers:
- url: "/"
security:
- bearer: []
...
paths:
  /admin/auth/:
    get:
      tags:
      - "登录API"
      summary: "首页"
      operationId: "index"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "string"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/captchaImage:
    get:
      tags:
      - "登录API"
      operationId: "getCaptchaImg"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaResponse"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Get captcha img"
  /admin/auth/get-async-routes:
    get:
      tags:
      - "登录API"
      summary: "获取异步路由"
      description: "兼容vue-pure-admin前端的路由获取接口"
      operationId: "getAsyncRoutes"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/RouterDTO"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/getConfig:
    get:
      tags:
      - "登录API"
      summary: "获取系统配置"
      operationId: "getConfig"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginConfigResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/getLoginUserInfo:
    get:
      tags:
      - "登录API"
      summary: "获取当前登录用户信息"
      operationId: "getLoginUserInfo"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentLoginUserResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/getRouters:
    get:
      tags:
      - "登录API"
      summary: "获取用户对应的菜单路由"
      description: "用于动态生成路由"
      operationId: "getRouters"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/RouterDTO"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/login:
    post:
      tags:
      - "登录API"
      summary: "登录"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/logout:
    post:
      tags:
      - "登录API"
      summary: "退出登录"
      operationId: "logout"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "boolean"
        "4XX":
          $ref: "#/components/responses/Error"
  /admin/auth/refresh-token:
    post:
      tags:
      - "登录API"
      summary: "刷新令牌"
      operationId: "refreshToken"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "4XX":
          $ref: "#/components/responses/Error"
...
  /admin/scheduler-job/log:
    get:
      tags:
      - "定时任务管理"
      summary: "查询任务执行日志"
      operationId: "logList"
      parameters:
      - name: "body"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/SchedulerLogListRequest"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPageResponseSchedulerLogResponse"
        "4XX":
          $ref: "#/components/responses/Error"
...
  /web/login:
    post:
      tags:
      - "web-auth-controller"
      operationId: "webLogin"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebLoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebLoginResponse"
        "4XX":
          $ref: "#/components/responses/Error"
      summary: "Web login"
...
components:
  schemas:
...
    AdminPageResponseSchedulerLogResponse:
      type: "object"
      properties:
        list:
          type: "array"
          items:
            $ref: "#/components/schemas/SchedulerLogResponse"
        total:
          type: "integer"
          format: "int64"
        pageSize:
          type: "integer"
          format: "int32"
        currentPage:
          type: "integer"
          format: "int32"
...
    CaptchaResponse:
      type: "object"
      properties:
        isCaptchaOn:
          type: "boolean"
        captchaCodeKey:
          type: "string"
        captchaCodeImg:
          type: "string"
...
    CurrentLoginUserResponse:
      type: "object"
      properties:
        userInfo:
          $ref: "#/components/schemas/UserResponse"
        roleKey:
          type: "string"
        permissions:
          uniqueItems: true
          type: "array"
          items:
            type: "string"
...
    DictionaryData:
      type: "object"
      properties:
        label:
          type: "string"
        value:
          type: "integer"
          format: "int32"
        cssTag:
          type: "string"
    ExtraIconDTO:
      type: "object"
      properties:
        svg:
          type: "boolean"
        name:
          type: "string"
...
    LoginConfigResponse:
      type: "object"
      properties:
        isCaptchaOn:
          type: "boolean"
        dictionary:
          type: "object"
          additionalProperties:
            type: "array"
            items:
              $ref: "#/components/schemas/DictionaryData"
...
    LoginRequest:
      required:
      - "password"
      - "username"
      type: "object"
      properties:
        username:
          minLength: 1
          type: "string"
        password:
          minLength: 1
          type: "string"
        captchaCode:
          type: "string"
        captchaCodeKey:
          type: "string"
      description: "登录信息"
    LoginResponse:
      type: "object"
      properties:
        avatar:
          type: "string"
        username:
          type: "string"
        nickname:
          type: "string"
        roles:
          type: "array"
          items:
            type: "string"
        permissions:
          type: "array"
          items:
            type: "string"
        accessToken:
          type: "string"
        refreshToken:
          type: "string"
        expires:
          type: "string"
...
    MetaDTO:
      type: "object"
      properties:
        title:
          type: "string"
        icon:
          type: "string"
        showLink:
          type: "boolean"
        showParent:
          type: "boolean"
        roles:
          type: "array"
          items:
            type: "string"
        auths:
          type: "array"
          items:
            type: "string"
        frameSrc:
          type: "string"
        isFrameSrcInternal:
          type: "boolean"
        rank:
          type: "integer"
          format: "int32"
        extraIcon:
          $ref: "#/components/schemas/ExtraIconDTO"
        keepAlive:
          type: "boolean"
        frameLoading:
          type: "boolean"
        transition:
          $ref: "#/components/schemas/TransitionDTO"
        hiddenTag:
          type: "boolean"
        dynamicLevel:
          type: "integer"
          format: "int32"
...
    RefreshTokenRequest:
      required:
      - "refreshToken"
      type: "object"
      properties:
        refreshToken:
          minLength: 1
          type: "string"
    RefreshTokenResponse:
      type: "object"
      properties:
        accessToken:
          type: "string"
        refreshToken:
          type: "string"
        expires:
          type: "string"
...
    RouterDTO:
      type: "object"
      properties:
        name:
          type: "string"
        path:
          type: "string"
        meta:
          $ref: "#/components/schemas/MetaDTO"
        children:
          type: "array"
          items:
            $ref: "#/components/schemas/RouterDTO"
...
    SchedulerLogListRequest:
      type: "object"
      properties:
        currentPage:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
        jobId:
          type: "integer"
          format: "int64"
    SchedulerLogResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        jobId:
          type: "integer"
          format: "int64"
        jobName:
          type: "string"
        jobGroup:
          type: "string"
        beanName:
          type: "string"
        methodName:
          type: "string"
        methodParams:
          type: "string"
        status:
          type: "integer"
          format: "int32"
        errorMessage:
          type: "string"
        durationMs:
          type: "integer"
          format: "int64"
        startedAt:
          type: "string"
          format: "date-time"
        finishedAt:
          type: "string"
          format: "date-time"
...
    TransitionDTO:
      type: "object"
      properties:
        name:
          type: "string"
        enterTransition:
          type: "string"
        leaveTransition:
          type: "string"
...
    UserResponse:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int64"
        roleId:
          type: "integer"
          format: "int64"
        roleName:
          type: "string"
        deptId:
          type: "integer"
          format: "int64"
        deptName:
          type: "string"
        username:
          type: "string"
        nickname:
          type: "string"
        userType:
          type: "integer"
          format: "int32"
        email:
          type: "string"
        phoneNumber:
          type: "string"
        sex:
          type: "integer"
          format: "int32"
        avatar:
          type: "string"
        status:
          type: "integer"
          format: "int32"
        loginIp:
          type: "string"
        loginDate:
          type: "string"
          format: "date-time"
        creatorId:
          type: "integer"
          format: "int64"
        creatorName:
          type: "string"
        createTime:
          type: "string"
          format: "date-time"
        updaterId:
          type: "integer"
          format: "int64"
        updaterName:
          type: "string"
        updateTime:
          type: "string"
          format: "date-time"
        remark:
          type: "string"
...
    WebLoginRequest:
      type: "object"
      properties:
        username:
          type: "string"
          minLength: 1
        password:
          type: "string"
          minLength: 1
      required:
      - "password"
      - "username"
    WebLoginResponse:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int64"
        username:
          type: "string"
        nickname:
          type: "string"
        avatar:
          type: "string"
        accessToken:
          type: "string"
        tokenName:
          type: "string"
        refreshToken:
          type: "string"
        expires:
          type: "string"
...
  securitySchemes:
    bearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  responses:
    Error:
      description: "Error response (RFC 9457 ProblemDetail)"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              type:
                type: "string"
              title:
                type: "string"
              status:
                type: "integer"
              detail:
                type: "string"
              instance:
                type: "string"